+++ stdc/net/Socket/TCPSocket/TCPSocket.h
@@ -9,7 +9,7 @@ typedef struct {
     Ptr                 (*new)      ();
     void                (*init)     (TCPSocketObject*);
     void                (*del)      (Ptr);
-    TCPSocketObject*    (*copy)     (TCPSocketObject*, int, MemoryObject*);
+    TCPSocketObject*    (*copy)     (TCPSocketObject*, long, MemoryObject*);
     
     // Server methods
     bool                (*bind)     (TCPSocketObject*, CStr, long);
