+++ stdc/net/Socket/TCPSocket/TCPSocket_private.h
@@ -7,7 +7,7 @@ static Ptr new_TCPSocket();
 static void init_TCPSocket(TCPSocketObject*);
 static void del_TCPSocket(Ptr);
 
-static TCPSocketObject* copy_TCPSocket(TCPSocketObject*, int, MemoryObject*);
+static TCPSocketObject* copy_TCPSocket(TCPSocketObject*, long, MemoryObject*);
 
 static bool bind_TCPSocket(TCPSocketObject*, CStr, long);
 static bool bindany_TCPSocket(TCPSocketObject*, long);
